requirements.txt
ultralytics
numpy
opencv-python
torch
torchvision
Pillow